+++ src/ashley/__init__.py
@@ -1,0 +1,5 @@
+"""Ashley is a django application that customizes and adds features to the django-machina forum."""
+
+
+# pylint: disable=invalid-name
+default_app_config = "ashley.apps.AshleyConfig"